encapsulate MNode manipulation inside MTree, update docs, change-list
