+++ server/src/main/java/org/apache/iotdb/db/metadata/MManager.java
@@ -20,7 +20,6 @@ package org.apache.iotdb.db.metadata;
 
 import org.apache.iotdb.db.concurrent.IoTDBThreadPoolFactory;
 import org.apache.iotdb.db.conf.IoTDBConfig;
-import org.apache.iotdb.db.conf.IoTDBConstant;
 import org.apache.iotdb.db.conf.IoTDBDescriptor;
 import org.apache.iotdb.db.engine.StorageEngine;
 import org.apache.iotdb.db.engine.cq.ContinuousQueryService;
@@ -50,7 +49,6 @@ import org.apache.iotdb.db.metadata.mnode.IMeasurementMNode;
 import org.apache.iotdb.db.metadata.mnode.IStorageGroupMNode;
 import org.apache.iotdb.db.metadata.mnode.MeasurementMNode;
 import org.apache.iotdb.db.metadata.mtree.MTree;
-import org.apache.iotdb.db.metadata.mtree.traverser.collector.CollectorTraverser;
 import org.apache.iotdb.db.metadata.path.MeasurementPath;
 import org.apache.iotdb.db.metadata.path.PartialPath;
 import org.apache.iotdb.db.metadata.tag.TagManager;
@@ -114,12 +112,10 @@ import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Deque;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.LinkedList;
@@ -2261,85 +2257,11 @@ public class MManager {
    * @return paths set
    */
   public Set<String> getPathsSetTemplate(String templateName) throws MetadataException {
-    Set<String> resSet = new HashSet<>();
-
-    CollectorTraverser<Set<String>> setTemplatePaths =
-        new CollectorTraverser<Set<String>>(
-            mtree.getNodeByPath(new PartialPath(IoTDBConstant.PATH_ROOT)),
-            new PartialPath(IoTDBConstant.PATH_ROOT).concatNode(MULTI_LEVEL_PATH_WILDCARD)) {
-          @Override
-          protected boolean processInternalMatchedMNode(IMNode node, int idx, int level)
-              throws MetadataException {
-            // will never get here, implement for placeholder
-            return false;
-          }
-
-          @Override
-          protected boolean processFullMatchedMNode(IMNode node, int idx, int level)
-              throws MetadataException {
-            // shall not traverse nodes inside template
-            if (!node.getPartialPath().equals(getCurrentPartialPath(node))) {
-              return true;
-            }
-
-            // if node not set template, go on traversing
-            if (node.getUpperTemplate() != null) {
-              // if set template, and equals to target or target for all, add to result
-              if (templateName.equals("")
-                  || templateName.equals(node.getUpperTemplate().getName())) {
-                resSet.add(node.getFullPath());
-              }
-              // descendants of the node cannot set another template, exit from this branch
-              return true;
-            }
-            return false;
-          }
-        };
-
-    setTemplatePaths.traverse();
-    return resSet;
+    return new HashSet<>(mtree.getPathsSetOnTemplate(templateName));
   }
 
   public Set<String> getPathsUsingTemplate(String templateName) throws MetadataException {
-    Set<String> result = new HashSet<>();
-
-    CollectorTraverser<Set<String>> usingTemplatePaths =
-        new CollectorTraverser<Set<String>>(
-            mtree.getNodeByPath(new PartialPath(IoTDBConstant.PATH_ROOT)),
-            new PartialPath(IoTDBConstant.PATH_ROOT).concatNode(MULTI_LEVEL_PATH_WILDCARD)) {
-          @Override
-          protected boolean processInternalMatchedMNode(IMNode node, int idx, int level)
-              throws MetadataException {
-            // will never get here, implement for placeholder
-            return false;
-          }
-
-          @Override
-          protected boolean processFullMatchedMNode(IMNode node, int idx, int level)
-              throws MetadataException {
-            // shall not traverse nodes inside template
-            if (!node.getPartialPath().equals(getCurrentPartialPath(node))) {
-              return true;
-            }
-
-            if (node.getUpperTemplate() != null) {
-              // this node and its descendants are set other template, exit from this branch
-              if (!templateName.equals("")
-                  && !templateName.equals(node.getUpperTemplate().getName())) {
-                return true;
-              }
-
-              // descendants of this node may be using template too
-              if (node.isUseTemplate()) {
-                result.add(node.getFullPath());
-              }
-            }
-            return false;
-          }
-        };
-
-    usingTemplatePaths.traverse();
-    return result;
+    return new HashSet<>(mtree.getPathsUsingTemplate(templateName));
   }
 
   public synchronized void dropSchemaTemplate(DropTemplatePlan plan) throws MetadataException {
@@ -2350,28 +2272,12 @@ public class MManager {
         throw new UndefinedTemplateException(templateName);
       }
 
-      // check whether template has been set
-      Deque<IMNode> nodeStack = new ArrayDeque<>();
-      nodeStack.push(mtree.getNodeByPath(new PartialPath(IoTDBConstant.PATH_ROOT)));
-
-      // DFT traverse on MTree
-      while (nodeStack.size() != 0) {
-        IMNode curNode = nodeStack.pop();
-        if (curNode.getUpperTemplate() != null) {
-          if (curNode.getUpperTemplate().getName().equals(templateName)) {
-            throw new MetadataException(
-                String.format(
-                    "Template [%s] has been set on [%s], cannot be dropped now.",
-                    templateName, curNode.getFullPath()));
-          }
-          // curNode set to other templates, cut this branch
-        }
-
-        // no template on curNode, push children to stack
-        for (IMNode child : curNode.getChildren().values()) {
-          nodeStack.push(child);
-        }
+      if (mtree.isTemplateSetOnMTree(templateName)) {
+        throw new MetadataException(
+            String.format(
+                "Template [%s] has been set on MTree, cannot be dropped now.", templateName));
       }
+
       templateManager.dropSchemaTemplate(plan);
       if (!isRecovering) {
         logWriter.dropSchemaTemplate(plan);
